+++ .eslintignore
@@ -853,9 +853,36 @@ packages/app-mobile/components/NoteBodyViewer/hooks/useOnResourceLongPress.js.ma
 packages/app-mobile/components/NoteBodyViewer/hooks/useSource.d.ts
 packages/app-mobile/components/NoteBodyViewer/hooks/useSource.js
 packages/app-mobile/components/NoteBodyViewer/hooks/useSource.js.map
-packages/app-mobile/components/NoteEditor/CodeMirror.d.ts
-packages/app-mobile/components/NoteEditor/CodeMirror.js
-packages/app-mobile/components/NoteEditor/CodeMirror.js.map
+packages/app-mobile/components/NoteEditor/CodeMirror/CodeMirror.d.ts
+packages/app-mobile/components/NoteEditor/CodeMirror/CodeMirror.js
+packages/app-mobile/components/NoteEditor/CodeMirror/CodeMirror.js.map
+packages/app-mobile/components/NoteEditor/CodeMirror/RegionSpec.d.ts
+packages/app-mobile/components/NoteEditor/CodeMirror/RegionSpec.js
+packages/app-mobile/components/NoteEditor/CodeMirror/RegionSpec.js.map
+packages/app-mobile/components/NoteEditor/CodeMirror/RegionSpec.test.d.ts
+packages/app-mobile/components/NoteEditor/CodeMirror/RegionSpec.test.js
+packages/app-mobile/components/NoteEditor/CodeMirror/RegionSpec.test.js.map
+packages/app-mobile/components/NoteEditor/CodeMirror/decorators.d.ts
+packages/app-mobile/components/NoteEditor/CodeMirror/decorators.js
+packages/app-mobile/components/NoteEditor/CodeMirror/decorators.js.map
+packages/app-mobile/components/NoteEditor/CodeMirror/languages.d.ts
+packages/app-mobile/components/NoteEditor/CodeMirror/languages.js
+packages/app-mobile/components/NoteEditor/CodeMirror/languages.js.map
+packages/app-mobile/components/NoteEditor/CodeMirror/markdownCommands.d.ts
+packages/app-mobile/components/NoteEditor/CodeMirror/markdownCommands.js
+packages/app-mobile/components/NoteEditor/CodeMirror/markdownCommands.js.map
+packages/app-mobile/components/NoteEditor/CodeMirror/markdownMathParser.d.ts
+packages/app-mobile/components/NoteEditor/CodeMirror/markdownMathParser.js
+packages/app-mobile/components/NoteEditor/CodeMirror/markdownMathParser.js.map
+packages/app-mobile/components/NoteEditor/CodeMirror/theme.d.ts
+packages/app-mobile/components/NoteEditor/CodeMirror/theme.js
+packages/app-mobile/components/NoteEditor/CodeMirror/theme.js.map
+packages/app-mobile/components/NoteEditor/CodeMirror/types.d.ts
+packages/app-mobile/components/NoteEditor/CodeMirror/types.js
+packages/app-mobile/components/NoteEditor/CodeMirror/types.js.map
+packages/app-mobile/components/NoteEditor/CodeMirror/webviewLogger.d.ts
+packages/app-mobile/components/NoteEditor/CodeMirror/webviewLogger.js
+packages/app-mobile/components/NoteEditor/CodeMirror/webviewLogger.js.map
 packages/app-mobile/components/NoteEditor/EditLinkDialog.d.ts
 packages/app-mobile/components/NoteEditor/EditLinkDialog.js
 packages/app-mobile/components/NoteEditor/EditLinkDialog.js.map
@@ -865,18 +892,9 @@ packages/app-mobile/components/NoteEditor/MarkdownToolbar.js.map
 packages/app-mobile/components/NoteEditor/NoteEditor.d.ts
 packages/app-mobile/components/NoteEditor/NoteEditor.js
 packages/app-mobile/components/NoteEditor/NoteEditor.js.map
-packages/app-mobile/components/NoteEditor/codeMirrorDecorators.d.ts
-packages/app-mobile/components/NoteEditor/codeMirrorDecorators.js
-packages/app-mobile/components/NoteEditor/codeMirrorDecorators.js.map
-packages/app-mobile/components/NoteEditor/codeMirrorLanguages.d.ts
-packages/app-mobile/components/NoteEditor/codeMirrorLanguages.js
-packages/app-mobile/components/NoteEditor/codeMirrorLanguages.js.map
-packages/app-mobile/components/NoteEditor/codeMirrorTheme.d.ts
-packages/app-mobile/components/NoteEditor/codeMirrorTheme.js
-packages/app-mobile/components/NoteEditor/codeMirrorTheme.js.map
-packages/app-mobile/components/NoteEditor/markdownMathParser.d.ts
-packages/app-mobile/components/NoteEditor/markdownMathParser.js
-packages/app-mobile/components/NoteEditor/markdownMathParser.js.map
+packages/app-mobile/components/NoteEditor/SelectionFormatting.d.ts
+packages/app-mobile/components/NoteEditor/SelectionFormatting.js
+packages/app-mobile/components/NoteEditor/SelectionFormatting.js.map
 packages/app-mobile/components/NoteEditor/types.d.ts
 packages/app-mobile/components/NoteEditor/types.js
 packages/app-mobile/components/NoteEditor/types.js.map
