+++ packages/app-mobile/.gitignore
@@ -63,5 +63,5 @@ buck-out/
 lib/csstojs/
 lib/rnInjectedJs/
 dist/
-components/NoteEditor/CodeMirror.bundle.js
-components/NoteEditor/CodeMirror.bundle.min.js
+components/NoteEditor/CodeMirror/CodeMirror.bundle.js
+components/NoteEditor/CodeMirror/CodeMirror.bundle.min.js
