+++ packages/app-mobile/components/NoteEditor/CodeMirror/theme.ts
