+++ packages/app-mobile/package.json
@@ -8,6 +8,7 @@
     "start": "react-native start --reset-cache",
     "build": "gulp build",
     "tsc": "tsc --project tsconfig.json",
+    "test": "jest",
     "watch": "tsc --watch --preserveWatchOutput --project tsconfig.json",
     "clean": "node tools/clean.js",
     "buildInjectedJs": "gulp buildInjectedJs",
@@ -87,6 +88,7 @@
     "@lezer/highlight": "^1.0.0",
     "@rollup/plugin-node-resolve": "^13.0.0",
     "@rollup/plugin-typescript": "^8.2.1",
+    "@types/jest": "^28.1.3",
     "@types/node": "^14.14.6",
     "@types/react": "^16.9.55",
     "@types/react-native": "^0.64.4",
@@ -94,6 +96,7 @@
     "execa": "^4.0.0",
     "fs-extra": "^8.1.0",
     "gulp": "^4.0.2",
+    "jest": "^28.1.1",
     "jetifier": "^1.6.5",
     "metro-react-native-babel-preset": "^0.66.2",
     "nodemon": "^2.0.12",
